recovery: Modernize data/infomanager to C++20 stdlib
- InfoManager: load settings with a single read + in-memory parse, save
  via one serialized write (on-disk format unchanged); Handle reads now
  convert to std::optional<T>
- DataManager: merge _SetProgress into SetProgress(fraction, scoped);
  persist flag is bool now; rename to CamelCase (OutputVersion,
  UpdateTimezoneEnvironment, SetDeviceId, kBackingFile, persist_/data_/
  consts_); drop using namespace std from the header
- Replace unsigned long long with uint64_t (partitions, twrpTar,
  Fixup_Time_On_Boot); read/write .info values via InfoManager's
  operator[] (Get_Restore_Size, backup_info)
- Rewrite TWFunc::Check_For_TwrpFolder on std::filesystem; pin
  Archive_Type underlying type to int

+++ partitionmanager.cpp
@@ -1204,7 +1204,7 @@ bool TWPartitionManager::Run_Backup(bool adbbackup) {
     if (part_settings.file_time == 0)
         part_settings.file_time = 1;
     int img_bps = static_cast<int>(part_settings.img_bytes) / static_cast<int>(part_settings.img_time);
-    unsigned long long file_bps = part_settings.file_bytes / static_cast<int>(part_settings.file_time);
+    uint64_t file_bps = part_settings.file_bytes / static_cast<int>(part_settings.file_time);
 
     if (part_settings.file_bytes != 0)
         gui_msg(Msg("avg_backup_fs=Average backup rate for file systems: {1}")(UnitConversion::FormatBytesPerSecond(file_bps)));
@@ -1222,7 +1222,7 @@ bool TWPartitionManager::Run_Backup(bool adbbackup) {
         actual_backup_size = part_settings.file_bytes + part_settings.img_bytes;
 
     int prev_img_bps = 0, use_compression = 0;
-    unsigned long long prev_file_bps = 0;
+    uint64_t prev_file_bps = 0;
     DataManager::GetValue(TW_BACKUP_AVG_IMG_RATE, prev_img_bps);
     img_bps += (prev_img_bps * 4);
     img_bps /= 5;
